chore: updates for compatibility with deadline-cloud-test-fixtures
